feat : 파일 구조 다시 조정 및 데이터 가져오기
