+++ package.json
@@ -16,6 +16,7 @@
     "dompurify": "^3.0.5",
     "env-cmd": "^10.1.0",
     "framer-motion": "^10.12.17",
+    "query-string": "^8.1.0",
     "quill": "^1.3.7",
     "quill-image-resize": "^3.0.9",
     "react": "^18.2.0",
