+++ src/Router.tsx
@@ -7,39 +7,39 @@ import KakaoLoginRedirect from "@pages/SocialLoginRedirect/SocialLoginRedirect";
 import MovieLog from "@pages/MovieLog/MovieLog";
 import EntireMovie from "@pages/EntireMovie/EntireMovie";
 import MovieTalk from "@pages/MovieTalk/MovieTalk";
-import SearchResult from "@pages/Search/SearchResult";
+import SearchResult from "@pages/SearchResult/SearchResult";
 import ReadReview from "@pages/ReadReview/ReadReview";
 import SingleMovie from "@pages/SingleMovie/SingleMovie";
 
 export const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-    children: [
-      {
-        path: "",
-        element: <Home />,
-      },
-      {
-        path: "write-review/:movieId",
-        element: <WriteReview />,
-      },
-      {
-        path: "/kakao-redirect/:accessToken/:refreshToken",
-        element: <KakaoLoginRedirect />,
-      },
-      { path: "entireMovie", element: <EntireMovie /> },
-      { path: "movietalk", element: <MovieTalk /> },
-      {
-        path: "movieLog/:userId",
-        element: <MovieLog />,
-      },
-      {
-        path: "review/:reviewId",
-        element: <ReadReview />,
-      },
-      { path: "/search-result", element: <SearchResult /> },
-    ],
-    errorElement: <NotFound />,
-  },
+    {
+        path: "/",
+        element: <App />,
+        children: [
+            {
+                path: "",
+                element: <Home />,
+            },
+            {
+                path: "write-review/:movieId",
+                element: <WriteReview />,
+            },
+            {
+                path: "/kakao-redirect/:accessToken/:refreshToken",
+                element: <KakaoLoginRedirect />,
+            },
+            { path: "entireMovie", element: <EntireMovie /> },
+            { path: "movietalk", element: <MovieTalk /> },
+            {
+                path: "movieLog/:userId",
+                element: <MovieLog />,
+            },
+            {
+                path: "review/:reviewId",
+                element: <ReadReview />,
+            },
+            { path: "/search-result", element: <SearchResult /> },
+        ],
+        errorElement: <NotFound />,
+    },
 ]);
