+++ src/components/Search/MovieResult/MovieResult.tsx
@@ -1,3 +1,15 @@
+import { SearchResultApi } from "@apis/searchResultApi";
+import { IMovieResult } from "@interfaces/movieResult";
+import { useQuery } from "@tanstack/react-query";
+import { useParams } from "react-router-dom";
+
 export default function MovieResult() {
-    return <>영화 결과 보여주기</>;
+    const { movieTitle } = useParams();
+
+    const { data: movieResult } = useQuery<IMovieResult>({
+        queryKey: ["movieResult"],
+        queryFn: async () => await SearchResultApi.getMovieResult(String(movieTitle)),
+    });
+
+    return <>{movieResult?.movieTitle}</>;
 }
