+++ src/components/Search/UserResult/UserResult.tsx
@@ -1,3 +1,3 @@
 export default function UserResult() {
-    return <>유저 결과 보여주기</>;
+    return <></>;
 }
