+++ src/pages/SearchResult/SearchResult.tsx
@@ -1,7 +1,4 @@
-import Movie from "@components/MovieLog/Movies/Movie/Movie";
 import Search from "@components/Search/Search";
-import MovieResult from "@components/Search/MovieResult/MovieResult";
-import UserResult from "@components/Search/UserResult/UserResult";
 import { Block } from "@styles/UI";
 
 export default function SearchResult() {
@@ -13,10 +10,8 @@ export default function SearchResult() {
                 </Block.RowBox>
                 <Block.Bar width="900px" height="1px" bgColor="gray" />
                 <Block.RowBox width="900px" height="30vw" bgColor="white" justifyContent="center" alignItems="center">
-                    <div> 검색 결과 나오는 부분</div>
-
-                    <MovieResult />
-                    <UserResult />
+                    {/* <MovieResult /> */}
+                    {/* <UserResult /> */}
                 </Block.RowBox>
             </Block.ColumnBox>
         </>
